requirements.txt
django==4.1.2
djangorestframework
django-cors-headers
djangorestframework-simplejwt
requests
django-modeltranslation
phonenumbers==8.13.26
django-phonenumber-field==7.2.0
djangorestframework-simplejwt==5.2.2
django-jazzmin==2.6.0